命名实体识别/requirement.txt
numpy==1.19.2
pandas==1.0.3
tensorflow-cpu==2.6.0
keras==2.6.0
protobuf==3.20.1
tensorflow-addons